test/requirements.txt
alabaster==1.0.0
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
docutils==0.21.2
flatbuffers==25.9.23
fonttools==4.60.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.0
opencv-python-headless==4.12.0.88
packaging==25.0
pbr==7.0.1
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
Pygments==2.19.2
PyMatting==1.1.14
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pyvista==0.46.3
referencing==0.36.2
rembg==2.0.67
requests==2.32.5
roman-numerals-py==3.1.0
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.16.2
scooby==0.10.2
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-autodoc-typehints==3.2.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-apidoc==0.6.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.14.0
tifffile==2025.9.30
tqdm==4.67.1
trimesh==4.8.3
typing_extensions==4.15.0
urllib3==2.5.0
vtk==9.5.2
